backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow>=10.0.0
easyocr==1.7.1
pytesseract==0.3.10
groq==0.4.1
supabase==2.3.4
cloudinary==1.38.0
python-dotenv==1.0.0
pydantic==2.5.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
backboard-sdk==0.1.0
pytest>=7.4.0
azure-ai-formrecognizer>=3.3.0
openpyxl>=3.1.0
xlrd>=2.0.1
